Updated Bedrock and OpenAI
