+++ src/models/llamaindex/llama-openai-model.ts
@@ -1,4 +1,4 @@
-import { OpenAI, Settings, ChatSummaryMemoryBuffer, ChatMessage, ALL_AVAILABLE_OPENAI_MODELS } from 'llamaindex'
+import { OpenAI, ChatMessage, ALL_AVAILABLE_OPENAI_MODELS } from 'llamaindex'
 import createDebugMessages from 'debug';
 import { BaseModel } from '../../interfaces/base-model.js';
 import { Chunk, ConversationHistory } from '../../global/types.js';
@@ -11,7 +11,6 @@ export class LlamaOpenAi extends BaseModel {
     private readonly modelName: string;
     private model: OpenAI;
     private dimensions: number;
-    private chatHistory: ChatSummaryMemoryBuffer;
 
     constructor(params?: { temperature?: number; modelName?: string; }) {
         super(params?.temperature ?? 0.1);
@@ -34,15 +33,9 @@ export class LlamaOpenAi extends BaseModel {
         this.model = new OpenAI({ 
             temperature: this.temperature, 
             model: this.modelName, 
-            maxTokens: this.dimensions, // CHANGE MAX TOKENS
-            azure: { // HOW TO SET FOR ONLY OPEN AI??
-                apiKey: process.env.AZURE_OPENAI_API_KEY,
-                apiVersion: process.env.AZURE_OPENAI_API_VERSION,
-                endpoint: `https://${process.env.AZURE_OPENAI_API_INSTANCE_NAME}.openai.azure.com` 
-            }
-        }); 
-        this.chatHistory = new ChatSummaryMemoryBuffer({ llm: this.model });
-        Settings.llm = this.model;
+            maxTokens: this.dimensions * 0.75, // CHANGE MAX TOKENS
+            apiKey: process.env.OPENAI_API_KEY,
+        });
     }
 
     protected async runQuery(
@@ -51,30 +44,13 @@ export class LlamaOpenAi extends BaseModel {
         supportingContext: Chunk[],
         pastConversations: ConversationHistory[]
     ): Promise<string> {
-        const contextSummary = supportingContext.map((chunk) => chunk.pageContent).join('; ');
-
-        this.chatHistory.put({ role: 'system', content: system });
-        this.chatHistory.put({ role: 'system', content: `Supporting context: ${contextSummary}` });
-
-        pastConversations.forEach((conversation) => {
-            if (conversation.sender === 'AI') {
-                this.chatHistory.put({ role: 'assistant', content: conversation.message });
-            } else if (conversation.sender === 'SYSTEM') {
-                this.chatHistory.put({ role: 'system', content: conversation.message });
-            } else {
-                this.chatHistory.put({ role: 'user', content: conversation.message });
-            }
-        });
-
-        // User's query
-        this.chatHistory.put({ role: 'user', content: userQuery });
-
-        this.debug('Executing query with prompt:', userQuery);
-        this.debug('Chat history:', this.chatHistory.getAllMessages());
-
-        const messages: ChatMessage[] = await this.chatHistory.getAllMessages();
-
-        const response = await this.model.chat({ messages });
+        const pastMessages: ChatMessage[] = this.generatePastMessagesLlama(
+            system,
+            supportingContext,
+            pastConversations,
+            userQuery,
+        );
+        const response = await this.model.chat({ messages: pastMessages });
         this.debug('Model response:', response);
 
         if (response.message && typeof response.message.content === 'string') {
